Rename main gbd package, adjust .gitignore
