+++ main/gbd_tool/import_data.py
@@ -1,7 +1,7 @@
 import csv
 import re
 
-from main.gbd.database import groups, benchmark_administration
+from main.gbd_tool.database import groups, benchmark_administration
 
 
 def exists(database, cat):
